Moved enums module to qudi.util namespace
Adapted imports for datastorage and datafitting modules.
