+++ qudi/hardware/microwave/mw_source_agilent.py
@@ -30,7 +30,7 @@ import numpy as np
 
 from qudi.core.configoption import ConfigOption
 from qudi.interface.microwave_interface import MicrowaveInterface, MicrowaveConstraints
-from qudi.core.enums import SamplingOutputMode, TriggerEdge
+from qudi.util.enums import SamplingOutputMode
 from qudi.util.mutex import Mutex
