+++ qudi/interface/microwave_interface.py
@@ -23,7 +23,7 @@ top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi
 from abc import abstractmethod
 
 from qudi.core.module import Base
-from qudi.core.enums import TriggerEdge, SamplingOutputMode
+from qudi.util.enums import SamplingOutputMode
 from qudi.util.helpers import in_range
