+++ qudi/util/enums.py
